Fix f32 to usize assignment

+++ src/ui/text.rs
@@ -14,7 +14,7 @@ impl Text {
     pub fn get_width(&self, size: f32) -> usize {
         let mut width = 0;
         for c in self.text.chars() {
-            width += self.font.font.metrics(c, size).advance_width;
+            width += self.font.font.metrics(c, size).advance_width as usize;
         }
         return width;
     }
